PWGEM/Dilepton: add MCUtilities.h for HFee
